move validateRegistryCommit definitions to git utils file
